requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask-Login==0.5.0
cryptography~=42.0.8
Pillow~=10.4.0
MarkupSafe==2.0.1

Werkzeug~=1.0.1
wtforms~=3.1.2
SQLAlchemy==1.3.24